+++ packages/@cdk-cosmos/core/src/galaxy/galaxy-core-import.ts
@@ -1,14 +1,16 @@
 import { Construct, Stack } from '@aws-cdk/core';
+import { Config } from '@cosmos-building-blocks/common';
+import { isCrossAccount } from '@cosmos-building-blocks/common/lib/utils';
 import { ICosmosCore } from '../cosmos/cosmos-core-stack';
 import { IGalaxyCore } from './galaxy-core-stack';
 import { BaseConstructProps, BaseConstruct } from '../components/base';
-import { isCrossAccount } from '../helpers/utils';
 
 export interface GalaxyCoreImportProps extends BaseConstructProps {
   cosmos: ICosmosCore;
 }
 
 export class GalaxyCoreImport extends BaseConstruct implements IGalaxyCore {
+  readonly config: Config;
   readonly cosmos: ICosmosCore;
   readonly cdkCrossAccountRoleStaticArn?: string;
 
@@ -22,6 +24,7 @@ export class GalaxyCoreImport extends BaseConstruct implements IGalaxyCore {
     const { cosmos } = props;
 
     this.cosmos = cosmos;
+    this.config = new Config(this, 'Config', id, this.cosmos.config);
 
     if (isCrossAccount(this, this.cosmos)) {
       const CdkCrossAccountRoleName = this.cosmos.singletonId('CdkCrossAccountRole');